modify Article List Component and make document page like StoryBoard
