+++ src/app/page.tsx
@@ -1,211 +1,25 @@
 "use client";
 
 import SearchBar from "@/components/searchBar";
-import BoardPreview from "@/components/BoardPreview/BoardPreview";
+import ArticleList from "@/components/ArticleList/ArticleList";
 import { useState } from "react";
-
-// 옵션 타입 정의
-interface UIOptions {
-  showWriter: boolean;
-  showBoard: boolean;
-  showProfile: boolean;
-  showHit: boolean;
-  showStatus: boolean;
-  showAttachment: boolean;
-  showRank: boolean;
-  showAnswerStatus: boolean;
-  showTimeAgo: boolean;
-}
-
-// 옵션 컨트롤 컴포넌트 타입 정의
-interface OptionControlProps {
-  options: UIOptions;
-  onChange: (option: string, value: boolean) => void;
-  containerWidth: number;
-  onWidthChange: (width: number) => void;
-  titleFontSize: string;
-  onFontSizeChange: (size: string) => void;
-  titleFontWeight: string;
-  onFontWeightChange: (weight: string) => void;
-}
-
-// 옵션 컨트롤 컴포넌트
-const OptionControl = ({ 
-  options, 
-  onChange, 
-  containerWidth, 
-  onWidthChange, 
-  titleFontSize, 
-  onFontSizeChange, 
-  titleFontWeight, 
-  onFontWeightChange 
-}: OptionControlProps) => {
-  return (
-    <div className="p-4 bg-white rounded-lg shadow-sm mb-4 border border-gray-200">
-      <h3 className="text-lg font-bold mb-3">UI 옵션 설정</h3>
-      
-      {/* 크기 조정 슬라이더 */}
-      <div className="mb-4 p-3 bg-gray-50 rounded-lg">
-        <label className="block text-sm font-medium mb-2">
-          컨테이너 너비: {containerWidth}px
-        </label>
-        <input
-          type="range"
-          min="300"
-          max="800"
-          value={containerWidth}
-          onChange={(e) => onWidthChange(parseInt(e.target.value))}
-          className="w-full h-2 bg-gray-200 rounded-lg appearance-none cursor-pointer"
-        />
-        <div className="flex justify-between text-xs text-gray-500 mt-1">
-          <span>300px</span>
-          <span>550px (기본)</span>
-          <span>800px</span>
-        </div>
-      </div>
-
-      {/* 폰트 크기 조정 드롭다운 */}
-      <div className="mb-4 p-3 bg-gray-50 rounded-lg">
-        <label className="block text-sm font-medium mb-2">
-          제목 폰트 크기: {titleFontSize}
-        </label>
-        <select
-          value={titleFontSize}
-          onChange={(e) => onFontSizeChange(e.target.value)}
-          className="w-full p-2 border border-gray-300 rounded-lg"
-        >
-          <option value="text-sm">Small</option>
-          <option value="text-base">Base</option>
-          <option value="text-lg">Large</option>
-          <option value="text-xl">Extra Large</option>
-          <option value="text-2xl">2XL</option>
-        </select>
-      </div>
-
-      {/* 폰트 굵기 조정 드롭다운 */}
-      <div className="mb-4 p-3 bg-gray-50 rounded-lg">
-        <label className="block text-sm font-medium mb-2">
-          제목 폰트 굵기: {titleFontWeight}
-        </label>
-        <select
-          value={titleFontWeight}
-          onChange={(e) => onFontWeightChange(e.target.value)}
-          className="w-full p-2 border border-gray-300 rounded-lg"
-        >
-          <option value="font-normal">Normal</option>
-          <option value="font-bold">Bold</option>
-          <option value="font-light">Light</option>
-          <option value="font-medium">Medium</option>
-          <option value="font-extrabold">Extra Bold</option>
-        </select>
-      </div>
-
-      <div className="grid grid-cols-3 gap-2">
-        {Object.entries(options).map(([key, value]) => (
-          <div key={key} className="flex items-center">
-            <input
-              type="checkbox"
-              id={key}
-              checked={value}
-              onChange={() => onChange(key, !value)}
-              className="mr-2"
-            />
-            <label htmlFor={key} className="text-sm">
-              {key.replace('show', '')}
-            </label>
-          </div>
-        ))}
-      </div>
-    </div>
-  );
-};
+import Link from "next/link";
+import Image from "next/image";
 
 export default function Home() {
   const [inputValue, setInputValue] = useState("");
-  const [containerWidth, setContainerWidth] = useState(550); // 기본 크기
-  const [titleFontSize, setTitleFontSize] = useState("text-base"); // 기본 폰트 크기
-  const [titleFontWeight, setTitleFontWeight] = useState("font-normal"); // 기본 폰트 굵기
-  
-  // UI 옵션 상태 관리
-  const [uiOptions, setUiOptions] = useState<UIOptions>({
-    showWriter: true,
-    showBoard: true,
-    showProfile: false,
-    showHit: true,
-    showStatus: true, 
-    showAttachment: true,
-    showRank: false,
-    showAnswerStatus: false,
-    showTimeAgo: true
-  });
-  
-  // UI 옵션 변경 핸들러
-  const handleOptionChange = (option: string, value: boolean) => {
-    setUiOptions(prev => ({
-      ...prev,
-      [option]: value
-    }));
-  };
 
-  // 테스트 데이터
-  const testPosts = [
-    { 
-      id: 1, 
-      title: "이것은 매우 긴 게시글 제목입니다. 실제로는 이보다 더 길 수도 있고 말줄임표(...) 처리가 잘 되는지 확인해보겠습니다", 
-      author: "개발자", 
-      timeAgo: "9분 전", 
-      likes: 50, 
-      dislikes: 2, 
-      comments: 15,
-      boardName: "개발 게시판",
-      hasAttachment: true,
-      attachmentType: 'image' as const,
-      hit: 230,
-      answered: true,
-      profileImage: "/assets/ServiceAra.svg"
-    },
-    { 
-      id: 2, 
-      title: "UI 옵션을 변경해보세요!", 
-      author: "디자이너", 
-      timeAgo: "20분 전", 
-      likes: 40, 
-      dislikes: 3, 
-      comments: 10,
-      boardName: "디자인 게시판",
-      hasAttachment: false,
-      hit: 185,
-      answered: false
-    },
-    { 
-      id: 3, 
-      title: "프로필 이미지와 순위 표시 테스트 - 이 제목도 상당히 길게 만들어서 말줄임 처리를 확인해보겠습니다", 
-      author: "테스터", 
-      timeAgo: "1시간 전", 
-      likes: 35, 
-      dislikes: 1, 
-      comments: 8,
-      boardName: "테스트 게시판",
-      hasAttachment: true,
-      attachmentType: 'both' as const,
-      hit: 142,
-      answered: true,
-      profileImage: "/assets/ServiceAra.svg"
-    },
-    { 
-      id: 4, 
-      title: "첨부파일 표시 기능 확인", 
-      author: "사용자", 
-      timeAgo: "2시간 전", 
-      likes: 25, 
-      dislikes: 0, 
-      comments: 5,
-      boardName: "일반 게시판",
-      hasAttachment: true,
-      attachmentType: 'file' as const,
-      hit: 98,
-      answered: false
-    }
+  // 홈페이지에서 사용할 Mock 데이터
+  const hotArticles = [
+    { id: 1, title: "SPARCS 정기 모집", author: "운영진", timeAgo: "1일 전", likes: 150, dislikes: 1, comments: 25, boardName: "공지", hasAttachment: true, hit: 1024 },
+    { id: 2, title: "딸기 파티에 초대합니다!", author: "딸기수호대", timeAgo: "5시간 전", likes: 99, dislikes: 5, comments: 32, boardName: "자유", hasAttachment: true, hit: 876 },
+    { id: 3, title: "새로운 아라 디자인 피드백 받습니다", author: "개발팀", timeAgo: "2일 전", likes: 85, dislikes: 2, comments: 41, boardName: "개발", hasAttachment: false, hit: 950 },
+  ];
+
+  const recentArticles = [
+    { id: 4, title: "오늘 점심 메뉴 추천", author: "점심봇", timeAgo: "5분 전", likes: 2, dislikes: 0, comments: 1, boardName: "자유", hasAttachment: false, hit: 15 },
+    { id: 5, title: "분실물 찾아가세요 (검은색 에어팟 프로)", author: "학생회", timeAgo: "30분 전", likes: 5, dislikes: 0, comments: 2, boardName: "공지", hasAttachment: true, hit: 55 },
+    { id: 6, title: "프로그래밍 스터디원 구합니다", author: "코딩고수", timeAgo: "1시간 전", likes: 10, dislikes: 0, comments: 3, boardName: "스터디", hasAttachment: false, hit: 88 },
   ];
 
   return (
@@ -218,38 +32,27 @@ export default function Home() {
           onChange={(e: React.ChangeEvent<HTMLInputElement>) => setInputValue(e.target.value)}
         />
       </div>
-      
-      {/* UI 옵션 컨트롤 */}
-      <div className="mb-6">
-        <OptionControl 
-          options={uiOptions} 
-          onChange={handleOptionChange}
-          containerWidth={containerWidth}
-          onWidthChange={setContainerWidth}
-          titleFontSize={titleFontSize}
-          onFontSizeChange={setTitleFontSize}
-          titleFontWeight={titleFontWeight}
-          onFontWeightChange={setTitleFontWeight}
-        />
-      </div>
-      
-      {/* 게시판 프리뷰 테스트 영역 */}
-      <div className="flex justify-center">
-        <div 
-          className="p-6 border border-gray-200 rounded-[20px] shadow-sm bg-white transition-all duration-300"
-          style={{ width: `${containerWidth + 48}px` }} // padding 24px * 2 = 48px 추가
-        >
-          <BoardPreview
-            boardTitle="🧪 게시판 컴포넌트 테스트"
-            posts={testPosts}
-            boardLink="/board/test"
-            containerWidth={containerWidth}
-            titleFontSize={titleFontSize}
-            titleFontWeight={titleFontWeight}
-            {...uiOptions}
-          />
-        </div>
-      </div>
+
+      {/* 실제 메인 페이지 컨텐츠 */}
+      <main className="flex flex-col items-center space-y-6 px-4">
+        {/* HOT 게시물 */}
+        <section className="w-full max-w-[550px] p-4 bg-white rounded-lg shadow-sm">
+          <Link href="/board/hot" className="flex items-center space-x-2 mb-[10px]">
+            <h2 className="text-[20px] font-semibold">HOT 게시물</h2>
+            <Image src="/Right_Chevron.svg" width={8.84} height={15} alt="arrow" />
+          </Link>
+          <ArticleList posts={hotArticles} showRank={true} showBoard={true} />
+        </section>
+
+        {/* 최신 글 */}
+        <section className="w-full max-w-[550px] p-4 bg-white rounded-lg shadow-sm">
+          <Link href="/board/recent" className="flex items-center space-x-2 mb-[10px]">
+            <h2 className="text-[20px] font-semibold">최신 글</h2>
+            <Image src="/Right_Chevron.svg" width={8.84} height={15} alt="arrow" />
+          </Link>
+          <ArticleList posts={recentArticles} showBoard={true} />
+        </section>
+      </main>
     </div>
   );
 }
